+++ eq-author/src/App/QuestionnaireDesignPage/MainNavigation/UpdateQuestionnaireSettingsModal/index.js
@@ -22,7 +22,7 @@ const UpdateQuestionnaireSettingsModal = ({
 UpdateQuestionnaireSettingsModal.propTypes = {
   onClose: PropTypes.func.isRequired,
   onUpdateQuestionnaire: PropTypes.func.isRequired,
-  onSubmit: PropTypes.func.isRequired,
+  onSubmit: PropTypes.func,
 };
 
 export default withUpdateQuestionnaire(UpdateQuestionnaireSettingsModal);
