+++ eq-author/src/App/QuestionnaireDesignPage/MainNavigation/index.js
@@ -1,7 +1,7 @@
-import React, { useState } from "react";
+import React, { useState, profiler } from "react";
 import styled from "styled-components";
 import PropTypes from "prop-types";
-import { flowRight } from "lodash";
+import { flowRight, get, find } from "lodash";
 
 import { withRouter } from "react-router-dom";
 import gql from "graphql-tag";
@@ -30,6 +30,13 @@ import shareIcon from "App/QuestionnaireDesignPage/MainNavigation/icons/sharing-
 import viewIcon from "App/QuestionnaireDesignPage/MainNavigation/icons/view-survey-icon.svg?inline";
 
 import UpdateQuestionnaireSettingsModal from "./UpdateQuestionnaireSettingsModal";
+
+import { useQuery } from "@apollo/react-hooks";
+import GET_ALL_ANSWERS from "../../qcodes/QCodesTable/graphql/getAllAnswers.graphql";
+import Error from "components/Error";
+import Loading from "components/Loading";
+
+import { organiseAnswers, flattenAnswers } from "../../../utils/getAllAnswersFlatMap"
 
 import {
   buildQcodesPath,
@@ -87,8 +94,60 @@ export const UnwrappedMainNavigation = props => {
     (questionnaire || {}).id
     }`;
 
+
+///////////////////////////////////////////////////////// WIP ///////////////////////////////////////////
+
+    const questionnaireId = (questionnaire || {}).id;
+
+    //could we use getAnswers from utils in API repo instead????
+
+    const { loading, error, data } = useQuery(GET_ALL_ANSWERS, {
+      variables: { input: { questionnaireId } },
+      fetchPolicy: "network-only",
+    });
+  
+    if (loading) {
+      return <Loading height="100%">Questionnaire answers loading…</Loading>;
+    }
+    if (error) {
+      return <Error>Oops! Something went wrong</Error>;
+    }
+
+    let emptyQCode;
+
+    if (data.questionnaire) {
+      const { sections } = get(data, "questionnaire", []);
+
+      console.log('sections', sections);
+
+      const { answers } = organiseAnswers(sections);
+
+      const flatten = flattenAnswers(answers);
+      console.log('flattened answers', flatten);
+
+      //todo - Rather just load emptyQCode with a bool
+      emptyQCode = find(flatten, obj => obj.qCode === "" || obj.qCode === null);
+      console.log('emptyQCode----------------', emptyQCode);
+    }
+
+    function onRenderCallback(
+      id, // the "id" prop of the Profiler tree that has just committed
+      phase, // either "mount" (if the tree just mounted) or "update" (if it re-rendered)
+      actualDuration, // time spent rendering the committed update
+      baseDuration, // estimated time to render the entire subtree without memoization
+      startTime, // when React began rendering this update
+      commitTime, // when React committed this update
+      interactions // the Set of interactions belonging to this update
+    ) {
+      console.log(id, phase, actualDuration, baseDuration, interactions)
+    }
+
+///////////////////////////////////////////////////////// WIP - end ///////////////////////////////////////////
+
+
   return (
     <>
+    <React.Profiler id="MainNavigation" onRender={onRenderCallback}>
       <StyledMainNavigation data-test="main-navigation">
         <Flex>
           <UtilityBtns>
@@ -179,7 +238,10 @@ export const UnwrappedMainNavigation = props => {
                   <IconText nav icon={qcodeIcon}>
                     QCodes
                   </IconText>
-                  {questionnaire.qCodeErrorCount > 0 ? <SmallBadge data-test="small-badge" /> : null}
+                  
+                  {/* {questionnaire.qCodeErrorCount > 0 ? <SmallBadge data-test="small-badge" /> : null} */}
+                  {emptyQCode ? <SmallBadge data-test="small-badge" /> : null}
+
                 </RouteButton>
                 {me && <UserProfile nav signOut left client={client} />}
               </ButtonGroup>
@@ -188,6 +250,7 @@ export const UnwrappedMainNavigation = props => {
         </Flex>
         {children}
       </StyledMainNavigation>
+      
       
       {questionnaire && (
         <>
@@ -198,6 +261,7 @@ export const UnwrappedMainNavigation = props => {
           />
         </>
       )}
+      </React.Profiler>
     </>
   );
 };
